+++ cli/BuildTwigFeedsCacheCommand.php
@@ -59,6 +59,7 @@ class BuildTwigFeedsCacheCommand extends ConsoleCommand
         } else {
             $config['cache_path'] = $config['locator']->findResource('cache://', true) . '/twigfeeds/';
         }
+        $config['blueprint_path'] = $config['locator']->findResource('user://plugins/twigfeeds/blueprints.yaml', true);
         return $config;
     }